Extract Pomodoro timer state into provider
